+++ packages/i18n/src/case.js
@@ -1,8 +1,6 @@
 import CaseMapper from 'ilib/lib/CaseMapper';
 
-// Assigned by initCaseMappers(), which runs during module initialization via
-// updateLocale() in glue.ts and again on every locale change
-let toLowerCaseMapper: CaseMapper, toUpperCaseMapper: CaseMapper;
+let toLowerCaseMapper, toUpperCaseMapper;
 
 /**
  * Set CaseMapper object references to ilib's current locale (its most recently set, by default)
@@ -21,7 +19,7 @@ function initCaseMappers () {
  * @param {String} inString String to convert to lower case
  * @returns {String} Lower-cased string
  */
-const toLowerCase = function (inString: string): string {
+const toLowerCase = function (inString) {
 	if (inString != null) {
 		return toLowerCaseMapper.map(inString.toString());
 	}
@@ -35,7 +33,7 @@ const toLowerCase = function (inString: string): string {
  * @param {String} inString String to convert to upper case
  * @returns {String} Upper-cased string
  */
-const toUpperCase = function (inString: string): string {
+const toUpperCase = function (inString) {
 	if (inString != null) {
 		return toUpperCaseMapper.map(inString.toString());
 	}
