+++ packages/i18n/src/glue.js
@@ -1,5 +1,7 @@
+/* global ILIB_ADDITIONAL_RESOURCES_PATH */
+
 /*
- * glue.ts - glue code to fit ilib into enact
+ * glue.js - glue code to fit ilib into enyo
  *
  * Copyright © 2013-2014 LG Electronics, Inc.
  *
@@ -36,4 +38,4 @@ if (typeof window === 'object' && typeof window.UILocale !== 'undefined') {
 
 // we go ahead and run this once during loading of iLib settings are valid
 // during the loads of later libraries.
-updateLocale(null);
+updateLocale(null, true);
