+++ packages/i18n/src/index.js
@@ -1,0 +1,3 @@
+import './glue';
+
+export default from 'ilib';
