--- packages/i18n/src/index.ts
@@ -1,3 +1,0 @@
-import './glue';
-
-export {default} from 'ilib';
